% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promedio_mes.R
\name{promedio_mes}
\alias{promedio_mes}
\title{Find the average speed sorted by month given a dataframe with MERRA-2 data}
\usage{
promedio_mes(promedioym)
}
\arguments{
\item{promedioym}{A dataframe created from a CSV file with MERRA-2 data}
}
\value{
A dataframe with the average speed sorted by month
}
\description{
Find the average speed sorted by month given a dataframe with MERRA-2 data
}
