% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promedio_ym.R
\name{promedio_ym}
\alias{promedio_ym}
\title{Find the average speed sorted by year and month given a dataframe with MERRA-2 data}
\usage{
promedio_ym(datos)
}
\arguments{
\item{datos}{A dataframe created from a CSV file with MERRA-2 data}
}
\value{
A dataframe with the average speed sorted by year and month
}
\description{
Find the average speed sorted by year and month given a dataframe with MERRA-2 data
}
