% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quitar_cols.R
\name{quitar_cols}
\alias{quitar_cols}
\title{Eliminate aditional variables from your dataframe}
\usage{
quitar_cols(datos)
}
\arguments{
\item{datos}{A dataframe created from a CSV file with MERRA-2 data}
}
\value{
Your original dataframe just with the variable Speed50
}
\description{
Eliminate aditional variables from your dataframe
}
